requirements.txt
aiogram==3.20.0.post0
APScheduler==3.10.4
SQLAlchemy==2.0.40
aiosqlite==0.21.0
httpx==0.28.1
beautifulsoup4==4.13.3
lxml==5.3.1
python-dotenv==1.0.1
playwright==1.52.0
PyYAML==6.0.2
orjson==3.10.16
uvloop==0.21.0; platform_system != "Windows"
